Refactored Store common tests
- Make use of t.Run() with an array of store types
- Run different stores types in parallel for a given test
- Remove `-v` from `go test` command since it does not help to
  see successful runs, and it has sometimes caused issues on Travis
- Attempt to remove `-p=1` to allow packages to be run in parallel.
  This may cause failures if the memory need running packages in
  parallel is too high (running `-race` is memory greedy), it will
  cause build failures (process will be killed by OS).
