+++ stores/filestore_sub_test.go
@@ -14,224 +14,6 @@ import (
 	"github.com/nats-io/nats-streaming-server/util"
 )
 
-func TestFSRecoverSubUpdatesForDeleteSubOK(t *testing.T) {
-	cleanupDatastore(t)
-	defer cleanupDatastore(t)
-
-	fs := createDefaultFileStore(t)
-	defer fs.Close()
-
-	cs := storeCreateChannel(t, fs, "foo")
-	// Store one sub for which we are going to store updates
-	// and then delete
-	sub1 := storeSub(t, cs, "foo")
-	// This one will stay and should be recovered
-	sub2 := storeSub(t, cs, "foo")
-
-	// Add several pending seq for sub1
-	storeSubPending(t, cs, "foo", sub1, 1, 2, 3)
-
-	// Delete sub
-	storeSubDelete(t, cs, "foo", sub1)
-
-	// Add more updates
-	storeSubPending(t, cs, "foo", sub1, 4, 5)
-	storeSubAck(t, cs, "foo", sub1, 1)
-
-	// Delete unexisting subs
-	storeSubDelete(t, cs, "foo", sub2+1, sub2+2, sub2+3)
-
-	// Close the store
-	fs.Close()
-
-	// Recovers now, should not have any error
-	limits := testDefaultStoreLimits
-	limits.MaxSubscriptions = 1
-	fs, state, err := newFileStore(t, defaultDataStore, &limits)
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-	defer fs.Close()
-
-	cs = getRecoveredChannel(t, state, "foo")
-	getRecoveredSubs(t, state, "foo", 1)
-	// Make sure the subs count was not messed-up by the fact
-	// that the store recovered delete requests for un-recovered
-	// subscriptions.
-	// Since we have set the limit of subs to 1, and we have
-	// recovered one, we should fail creating a new one.
-	sub := &spb.SubState{
-		ClientID:      "me",
-		Inbox:         nuidGen.Next(),
-		AckInbox:      nuidGen.Next(),
-		AckWaitInSecs: 10,
-	}
-	if err := cs.Subs.CreateSub(sub); err == nil || err != ErrTooManySubs {
-		t.Fatalf("Should have failed creating a sub, got %v", err)
-	}
-}
-
-func TestFSNoSubIdCollisionAfterRecovery(t *testing.T) {
-	cleanupDatastore(t)
-	defer cleanupDatastore(t)
-
-	fs := createDefaultFileStore(t)
-	defer fs.Close()
-
-	cs := storeCreateChannel(t, fs, "foo")
-	// Store a subscription.
-	sub1 := storeSub(t, cs, "foo")
-
-	// Close the store
-	fs.Close()
-
-	// Recovers now
-	fs, state := openDefaultFileStore(t)
-	defer fs.Close()
-	cs = getRecoveredChannel(t, state, "foo")
-	getRecoveredSubs(t, state, "foo", 1)
-	// Store new subscription
-	sub2 := storeSub(t, cs, "foo")
-
-	if sub2 <= sub1 {
-		t.Fatalf("Invalid subscription id after recovery, should be at leat %v, got %v", sub1+1, sub2)
-	}
-
-	// Store a delete subscription with higher ID and make sure
-	// we use something higher on restart
-	delSub := uint64(sub1 + 10)
-	storeSubDelete(t, cs, "foo", delSub)
-
-	// Close the store
-	fs.Close()
-
-	// Recovers now
-	fs, state = openDefaultFileStore(t)
-	defer fs.Close()
-	cs = getRecoveredChannel(t, state, "foo")
-	// sub1 & sub2 should be recovered
-	getRecoveredSubs(t, state, "foo", 2)
-
-	// Store new subscription
-	sub3 := storeSub(t, cs, "foo")
-
-	if sub3 <= sub1 || sub3 <= delSub {
-		t.Fatalf("Invalid subscription id after recovery, should be at leat %v, got %v", delSub+1, sub3)
-	}
-}
-
-func TestFSSubLastSentCorrectOnRecovery(t *testing.T) {
-	cleanupDatastore(t)
-	defer cleanupDatastore(t)
-
-	fs := createDefaultFileStore(t)
-	defer fs.Close()
-
-	cs := storeCreateChannel(t, fs, "foo")
-	// Store a subscription.
-	subID := storeSub(t, cs, "foo")
-
-	// A message
-	msg := []byte("hello")
-
-	// Store msg seq 1 and 2
-	m1 := storeMsg(t, cs, "foo", msg)
-	m2 := storeMsg(t, cs, "foo", msg)
-
-	// Store m1 and m2 for this subscription, then m1 again.
-	storeSubPending(t, cs, "foo", subID, m1.Sequence, m2.Sequence, m1.Sequence)
-
-	// Restart server
-	fs.Close()
-	fs, state := openDefaultFileStore(t)
-	defer fs.Close()
-	subs := getRecoveredSubs(t, state, "foo", 1)
-	sub := subs[0]
-	// Check that sub's last seq is m2.Sequence
-	if sub.Sub.LastSent != m2.Sequence {
-		t.Fatalf("Expected LastSent to be %v, got %v", m2.Sequence, sub.Sub.LastSent)
-	}
-}
-
-func TestFSUpdatedSub(t *testing.T) {
-	cleanupDatastore(t)
-	defer cleanupDatastore(t)
-
-	fs := createDefaultFileStore(t)
-	defer fs.Close()
-
-	cs := storeCreateChannel(t, fs, "foo")
-	// Creeate a subscription.
-	subID := storeSub(t, cs, "foo")
-
-	// A message
-	msg := []byte("hello")
-
-	// Store msg seq 1 and 2
-	m1 := storeMsg(t, cs, "foo", msg)
-	m2 := storeMsg(t, cs, "foo", msg)
-	m3 := storeMsg(t, cs, "foo", msg)
-
-	// Store m1 and m2 for this subscription
-	storeSubPending(t, cs, "foo", subID, m1.Sequence, m2.Sequence)
-
-	// Update the subscription
-	ss := cs.Subs
-	updatedSub := &spb.SubState{
-		ID:            subID,
-		ClientID:      "me",
-		Inbox:         nuidGen.Next(),
-		AckInbox:      "newAckInbox",
-		AckWaitInSecs: 10,
-	}
-	if err := ss.UpdateSub(updatedSub); err != nil {
-		t.Fatalf("Error updating subscription: %v", err)
-	}
-	// Store m3 for this subscription
-	storeSubPending(t, cs, "foo", subID, m3.Sequence)
-
-	// Store a subscription with update only, should be recovered
-	subWithoutNew := &spb.SubState{
-		ID:            subID + 1,
-		ClientID:      "me",
-		Inbox:         nuidGen.Next(),
-		AckInbox:      nuidGen.Next(),
-		AckWaitInSecs: 10,
-	}
-	if err := ss.UpdateSub(subWithoutNew); err != nil {
-		t.Fatalf("Error updating subscription: %v", err)
-	}
-
-	// Restart server
-	fs.Close()
-	fs, state := openDefaultFileStore(t)
-	defer fs.Close()
-	subs := getRecoveredSubs(t, state, "foo", 2)
-	// Subscriptions are recovered from a map, and then returned as an array.
-	// There is no guarantee that we get them in the order they were persisted.
-	for _, s := range subs {
-		if s.Sub.ID == subID {
-			// Check that sub's last seq is m3.Sequence
-			if s.Sub.LastSent != m3.Sequence {
-				t.Fatalf("Expected LastSent to be %v, got %v", m3.Sequence, s.Sub.LastSent)
-			}
-			// Update lastSent since we know it is correct.
-			updatedSub.LastSent = m3.Sequence
-			// Now compare that what we recovered is same that we used to update.
-			if !reflect.DeepEqual(*s.Sub, *updatedSub) {
-				t.Fatalf("Expected subscription to be %v, got %v", updatedSub, s.Sub)
-			}
-		} else if s.Sub.ID == subID+1 {
-			// Compare that what we recovered is same that we used to update.
-			if !reflect.DeepEqual(*s.Sub, *subWithoutNew) {
-				t.Fatalf("Expected subscription to be %v, got %v", subWithoutNew, s.Sub)
-			}
-		} else {
-			t.Fatalf("Unexpected subscription ID: %v", s.Sub.ID)
-		}
-	}
-}
-
 func TestFSBadSubFile(t *testing.T) {
 	cleanupDatastore(t)
 	defer cleanupDatastore(t)
